fix(NFM-1216): resolve remaining CI failures — mypy + ReviewAuthGuard tests
- Remove 4 unused type:ignore comments in ontology_sync.py
- Cast float() in kg_re.py _compute_relation_confidence return
- Cast ResolutionStrategy enum in fusion_pipeline.py
- Guard None whereclause in ontology.py search endpoint
- Fix ShortestPathResponse field names (from_node→from_, to_node→to)
- Fix ReviewAuthGuard test: pending promise mock + correct loading assertion

+++ apps/api/src/nfm_db/services/kg_re.py
@@ -186,7 +186,7 @@ class RelationExtractor:
         """Compute relation confidence as geometric mean of entity confidences."""
         if source_confidence <= 0 or target_confidence <= 0:
             return 0.0
-        return (source_confidence * target_confidence) ** 0.5
+        return float((source_confidence * target_confidence) ** 0.5)
 
     @staticmethod
     def _build_relation_properties(
